healthcare_kiosk/backend/auth.py
# backend/auth.py

from fastapi import APIRouter, HTTPException, Depends, status
from fastapi.security import OAuth2PasswordBearer, OAuth2PasswordRequestForm
from sqlalchemy.orm import Session
from pydantic import BaseModel
from passlib.context import CryptContext
from datetime import datetime, timedelta
from jose import JWTError, jwt

from database import models, crud, database

# JWT Config
SECRET_KEY = "your-very-secret-key"  # 🔐 Replace with secure env variable
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_MINUTES = 60

# Password hashing
pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")

# OAuth2 setup
oauth2_scheme = OAuth2PasswordBearer(tokenUrl="/auth/token")

# FastAPI router
router = APIRouter(prefix="/auth", tags=["Authentication"])


# ----------------------------- SCHEMAS ----------------------------- #

class UserCreate(BaseModel):
    username: str
    password: str
    email: str
    full_name: str | None = None

class Token(BaseModel):
    access_token: str
    token_type: str

# -------------------------- UTILITY FUNCTIONS ---------------------- #

def get_password_hash(password: str) -> str:
    return pwd_context.hash(password)

def verify_password(plain_password: str, hashed_password: str) -> bool:
    return pwd_context.verify(plain_password, hashed_password)

def create_access_token(data: dict, expires_delta: timedelta = None):
    to_encode = data.copy()
    expire = datetime.utcnow() + (expires_delta or timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES))
    to_encode.update({"exp": expire})
    return jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)

def get_user(db: Session, username: str):
    return crud.get_user_by_username(db, username)

def authenticate_user(db: Session, username: str, password: str):
    user = get_user(db, username)
    if not user or not verify_password(password, user.hashed_password):
        return False
    return user

# --------------------------- ROUTES ----------------------------- #

@router.post("/register", status_code=201)
def register(user: UserCreate, db: Session = Depends(database.get_db)):
    existing_user = crud.get_user_by_username(db, user.username)
    if existing_user:
        raise HTTPException(status_code=400, detail="Username already exists.")
    
    hashed_pw = get_password_hash(user.password)
    crud.create_user(db, user.username, hashed_pw, user.email, user.full_name or "")
    return {"msg": "User registered successfully."}

@router.post("/token", response_model=Token)
def login(form_data: OAuth2PasswordRequestForm = Depends(), db: Session = Depends(database.get_db)):
    user = authenticate_user(db, form_data.username, form_data.password)
    if not user:
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Invalid username or password",
            headers={"WWW-Authenticate": "Bearer"},
        )
    token = create_access_token(data={"sub": user.username})
    return {"access_token": token, "token_type": "bearer"}

class UserOut(BaseModel):
    username: str
    email: str
    full_name: str | None = None
    created_at: datetime

@router.get("/me", response_model=UserOut)
def get_current_user(token: str = Depends(oauth2_scheme), db: Session = Depends(database.get_db)):
    credentials_exception = HTTPException(
        status_code=status.HTTP_401_UNAUTHORIZED,
        detail="Invalid token",
        headers={"WWW-Authenticate": "Bearer"},
    )
    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
        username = payload.get("sub")
        if username is None:
            raise credentials_exception
        user = get_user(db, username)
        if user is None:
            raise credentials_exception
        return UserOut(
            username=user.username,
            email=user.email,
            full_name=user.full_name,
            created_at=user.created_at
        )
    except JWTError:
        raise credentials_exception
