healthcare_kiosk/requirements.txt
# Core Web Frameworks
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1

# Database & ORM
sqlalchemy==2.0.23
alembic==1.13.1

# Data Validation & Serialization
pydantic==2.5.0
python-multipart==0.0.6

# HTTP Client
requests==2.31.0

# Computer Vision & Face Recognition
opencv-python==4.8.1.78
face-recognition==1.3.0
dlib==19.24.0

# Image Processing
Pillow==10.1.0
numpy==1.26.4

# Machine Learning & AI
scikit-learn==1.3.2
pandas==2.1.3

# Natural Language Processing
rapidfuzz==3.5.2

# Voice Processing
SpeechRecognition==3.10.0
pyttsx3==2.90
pyaudio==0.2.11

# Windows Audio Support (Windows only)
pywin32==306; sys_platform == "win32"

# Utilities
python-dotenv==1.0.0
pathlib2==2.3.7
typing-extensions==4.8.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Logging & Monitoring
loguru==0.7.2

# Security
cryptography==41.0.7
passlib==1.7.4

# Configuration Management
python-jose==3.3.0

# Data Processing
openpyxl==3.1.2
xlsxwriter==3.1.9

# Time & Date Handling
python-dateutil==2.8.2

# Environment & System
psutil==5.9.6

# Additional Dependencies for Production
gunicorn==21.2.0
redis==5.0.1

# Optional: For enhanced AI capabilities
# torch==2.1.0
# transformers==4.35.0
# spacy==3.7.2

# Optional: For cloud deployment
# boto3==1.34.0
# azure-storage-blob==12.19.0
